GCP_BackEnd/requirements.txt
google-cloud-speech>=1.2.0
google-cloud-translate>=2.0.0
srt>=3.0.0
flask
